Unify prevout tests fixtures.
