+++ internal/application/testdata/ark_prevout_fetcher.json
@@ -2,33 +2,45 @@
   "valid": [
     {
       "name": "intent psbt",
-      "packet": "intent_psbt"
+      "psbt": "cHNidP8BAFwCAAAAAhOE4SQHNO1XbzZRMJ8tdpxOV7qnMbyjW6tUKlwMdhBmAAAAAAAAAAAARna3pPtDQbkHL+BGMzxajaheE/eerfBEgUIwtn+pjSgBAAAAAAAAAAAAAAAAAAABAQroAwAAAAAAAAFRCt5wcmV2YXJrdHg9AQAAAAEBAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAECAAAAAAAAAAFRAAAAAAABAQroAwAAAAAAAAFRCt5wcmV2YXJrdHg9AQAAAAECAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAEDAAAAAAAAAAFRAAAAAAA=",
+      "checkpoints": []
     },
     {
       "name": "ark tx psbt",
-      "packet": "ark_tx_psbt"
+      "psbt": "cHNidP8BADMCAAAAAcUDtmZZcUzDxYQ52DaK2vJwh+4LxvM1ax0fSiqO7KLbAAAAAAAAAAAAAAAAAAAAAQEK6AMAAAAAAAABUQrecHJldmFya3R4PQEAAAABAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABAgAAAAAAAAABUQAAAAAA",
+      "checkpoints": [
+        "cHNidP8BAD0CAAAAAROE4SQHNO1XbzZRMJ8tdpxOV7qnMbyjW6tUKlwMdhBmAAAAAAAAAAAAAQEAAAAAAAAAAVEAAAAAAAEBCugDAAAAAAAAAVEAAA=="
+      ]
     }
   ],
   "invalid": [
     {
       "name": "mismatched prevout hash",
-      "packet": "mismatched_prevout_hash",
-      "expectedErr": "prevout tx hash mismatch"
+      "psbt": "cHNidP8BADMCAAAAAQkJCQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAQEK6AMAAAAAAAABUQrecHJldmFya3R4PQEAAAABAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABAgAAAAAAAAABUQAAAAAA",
+      "checkpoints": [],
+      "errorContains": "prevout tx hash mismatch"
     },
     {
       "name": "duplicate prevout psbt fields",
-      "packet": "duplicate_prevout_psbt_fields",
-      "decodeErrorContains": "duplicate key"
+      "psbt": "cHNidP8BADMCAAAAAROE4SQHNO1XbzZRMJ8tdpxOV7qnMbyjW6tUKlwMdhBmAAAAAAAAAAAAAAAAAAAAAQEK6AMAAAAAAAABUQrecHJldmFya3R4PQEAAAABAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABAgAAAAAAAAABUQAAAAAK3nByZXZhcmt0eD0BAAAAAQEAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAQIAAAAAAAAAAVEAAAAAAA==",
+      "checkpoints": [],
+      "errorContains": "duplicate key"
     },
     {
       "name": "checkpoint source tx mismatch",
-      "packet": "checkpoint_source_tx_mismatch",
-      "expectedErr": "prevout tx hash mismatch"
+      "psbt": "cHNidP8BADMCAAAAAcUDtmZZcUzDxYQ52DaK2vJwh+4LxvM1ax0fSiqO7KLbAAAAAAAAAAAAAAAAAAAAAQEK6AMAAAAAAAABUQrecHJldmFya3R4PQEAAAABAgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABAwAAAAAAAAABUQAAAAAA",
+      "checkpoints": [
+        "cHNidP8BAD0CAAAAAROE4SQHNO1XbzZRMJ8tdpxOV7qnMbyjW6tUKlwMdhBmAAAAAAAAAAAAAQEAAAAAAAAAAVEAAAAAAAEBCugDAAAAAAAAAVEAAA=="
+      ],
+      "errorContains": "prevout tx hash mismatch"
     },
     {
       "name": "checkpoint prevout output index out of range",
-      "packet": "checkpoint_prevout_output_index_out_of_range",
-      "expectedErr": "prevout tx output index out of range"
+      "psbt": "cHNidP8BADMCAAAAAR+RFHsaCIEntrx42/lL6GnszBogjc+D+wWbOycY9pN7AAAAAAAAAAAAAAAAAAAAAQEK6AMAAAAAAAABUQrecHJldmFya3R4PQEAAAABAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABAgAAAAAAAAABUQAAAAAA",
+      "checkpoints": [
+        "cHNidP8BAD0CAAAAAROE4SQHNO1XbzZRMJ8tdpxOV7qnMbyjW6tUKlwMdhBmKgAAAAAAAAAAAQEAAAAAAAAAAVEAAAAAAAEBCugDAAAAAAAAAVEAAA=="
+      ],
+      "errorContains": "prevout tx output index out of range"
     }
   ]
 }
