--- internal/application/testdata/psbt_packets.json
@@ -1,34 +1,0 @@
-{
-  "packets": {
-    "intent_psbt": {
-      "psbt": "cHNidP8BAFwCAAAAAhOE4SQHNO1XbzZRMJ8tdpxOV7qnMbyjW6tUKlwMdhBmAAAAAAAAAAAARna3pPtDQbkHL+BGMzxajaheE/eerfBEgUIwtn+pjSgBAAAAAAAAAAAAAAAAAAABAQroAwAAAAAAAAFRCt5wcmV2YXJrdHg9AQAAAAEBAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAECAAAAAAAAAAFRAAAAAAABAQroAwAAAAAAAAFRCt5wcmV2YXJrdHg9AQAAAAECAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAEDAAAAAAAAAAFRAAAAAAA=",
-      "checkpoints": []
-    },
-    "ark_tx_psbt": {
-      "psbt": "cHNidP8BADMCAAAAAcUDtmZZcUzDxYQ52DaK2vJwh+4LxvM1ax0fSiqO7KLbAAAAAAAAAAAAAAAAAAAAAQEK6AMAAAAAAAABUQrecHJldmFya3R4PQEAAAABAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABAgAAAAAAAAABUQAAAAAA",
-      "checkpoints": [
-        "cHNidP8BAD0CAAAAAROE4SQHNO1XbzZRMJ8tdpxOV7qnMbyjW6tUKlwMdhBmAAAAAAAAAAAAAQEAAAAAAAAAAVEAAAAAAAEBCugDAAAAAAAAAVEAAA=="
-      ]
-    },
-    "mismatched_prevout_hash": {
-      "psbt": "cHNidP8BADMCAAAAAQkJCQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAQEK6AMAAAAAAAABUQrecHJldmFya3R4PQEAAAABAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABAgAAAAAAAAABUQAAAAAA",
-      "checkpoints": []
-    },
-    "duplicate_prevout_psbt_fields": {
-      "psbt": "cHNidP8BADMCAAAAAROE4SQHNO1XbzZRMJ8tdpxOV7qnMbyjW6tUKlwMdhBmAAAAAAAAAAAAAAAAAAAAAQEK6AMAAAAAAAABUQrecHJldmFya3R4PQEAAAABAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABAgAAAAAAAAABUQAAAAAK3nByZXZhcmt0eD0BAAAAAQEAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAQIAAAAAAAAAAVEAAAAAAA==",
-      "checkpoints": []
-    },
-    "checkpoint_source_tx_mismatch": {
-      "psbt": "cHNidP8BADMCAAAAAcUDtmZZcUzDxYQ52DaK2vJwh+4LxvM1ax0fSiqO7KLbAAAAAAAAAAAAAAAAAAAAAQEK6AMAAAAAAAABUQrecHJldmFya3R4PQEAAAABAgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABAwAAAAAAAAABUQAAAAAA",
-      "checkpoints": [
-        "cHNidP8BAD0CAAAAAROE4SQHNO1XbzZRMJ8tdpxOV7qnMbyjW6tUKlwMdhBmAAAAAAAAAAAAAQEAAAAAAAAAAVEAAAAAAAEBCugDAAAAAAAAAVEAAA=="
-      ]
-    },
-    "checkpoint_prevout_output_index_out_of_range": {
-      "psbt": "cHNidP8BADMCAAAAAR+RFHsaCIEntrx42/lL6GnszBogjc+D+wWbOycY9pN7AAAAAAAAAAAAAAAAAAAAAQEK6AMAAAAAAAABUQrecHJldmFya3R4PQEAAAABAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABAgAAAAAAAAABUQAAAAAA",
-      "checkpoints": [
-        "cHNidP8BAD0CAAAAAROE4SQHNO1XbzZRMJ8tdpxOV7qnMbyjW6tUKlwMdhBmKgAAAAAAAAAAAQEAAAAAAAAAAVEAAAAAAAEBCugDAAAAAAAAAVEAAA=="
-      ]
-    }
-  }
-}
